cleaning up schemas
