+++ packages/common/src/repo/index.ts
@@ -2,4 +2,3 @@ export * from './repo'
 export * from './tid'
 export * from './mst'
 export * from './types'
-export * as delta from './delta'
